% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05__make_nested_set.r
\name{bl_distance_ns}
\alias{bl_distance_ns}
\title{bl_distance_ns}
\usage{
bl_distance_ns(tipa, tipb, tree, ns)
}
\arguments{
\item{tipa}{string. Name of a tip in tree.}

\item{tipb}{string. Name of another tip in tree.}

\item{tree}{phylo object. Tree containing all unique species in x as tips. May
contain tips that are not in x.}

\item{ns}{matrix. Nested-set matrix for tree; use make_nested_set(tree).}
}
\value{
Distance between tipa and tipb.
}
\description{
Calculates branch-length distance between tipa and tipb in a phylogenetic tree
using nested-set optomization. Requires a pre-calculated nested-set.
}
\examples{
  library(ape)
  example_tree <- ape::read.tree(text=" ((((a:1,b:1):1,c:2):1,d:3):1,(e:1,f:1):3);")
  plot(example_tree); axis(side=1)
  example_ns <- make_nested_set(example_tree)
  bl_distance_ns("a", "c", example_tree, example_ns) # should be 4
  bl_distance_ns("a", "f", example_tree, example_ns) # should be 8
  bl_distance_ns("d", "c", example_tree, example_ns) # should be 6
  

}
\author{
John L. Darcy
}
