% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daughter_seeds.r
\name{daughter_seeds}
\alias{daughter_seeds}
\title{daughter_seeds}
\usage{
daughter_seeds(n, s = 12345)
}
\arguments{
\item{n}{integer. Number of daughter seeds to make.}

\item{s}{integer. A seed (DEFAULT: 12345).}
}
\value{
vector of length n containing integer seeds.
}
\description{
Makes n daughter seeds from seed s. This is useful for processes one wishes to be
deterministic, but may not be executed in the same order every time.
}
\author{
John L. Darcy
}
