% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90__matrix_functions.r
\name{distcalc}
\alias{distcalc}
\title{geo_distcalc}
\usage{
distcalc(lat, lng, sampIDs = NULL)
}
\arguments{
\item{lat}{Numeric vector. Latitudes in decimal degree format.}

\item{lng}{Numeric vector. Longitudes in decimal degree format.}

\item{sampIDs}{Character vector. Sample identifiers. Only required
if output dist should have names associated.}
}
\value{
matrix containing all pairwise geographic distances in km.
}
\description{
Calculates pairwise geographic distance between locations on earth. Just a
convenient wrapper for fields::rdist.earth.
}
\examples{
data(endophyte)
geo_dists <- distcalc(metadata$Lat, metadata$Lon, metadata$SampleID)
all(rownames(geo_dists) == metadata$SampleID)

}
\author{
John L. Darcy
}
