% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40__env_spec_sim.r
\name{env_spec_sim}
\alias{env_spec_sim}
\title{env_spec_sim}
\usage{
env_spec_sim(sdev, ideal, ideal2 = 0, ideal3 = 0, n_ideal = 1, env,
  n_obs, up = 0, oceanp = 0, n_cores = 2, seed = 1234567)
}
\arguments{
\item{sdev}{numeric vector. Standard deviation of the probability distribution
P(species), in the same units of env. Low values mean that the species
is found acrosss only a narrow range of env, i.e. specificity. High values 
mean that the species is found across a wide range of env, i.e. 
cosmopolitanism. Multiple values can be input in order to simulate a range
of specificities simultaneously. Can be length 1 or n.}

\item{ideal}{numeric vector. Value of env that is ideal for the simulated
species. This is the mode of the probability distribution P(species).
Can be length 1 or n.}

\item{ideal2}{numeric vector. Value of env that is the second ideal for the 
simulated species. Only used if n_ideal >= 2. This is the second mode of the
probability distribution P(species). Can be length 1 or n.}

\item{ideal3}{numeric vector. Value of env that is the third ideal for the 
simulated species. Only used if n_ideal = 3. This is the third mode of the
probability distribution P(species). Can be length 1 or n.}

\item{env}{numeric vector. Real or fake environmental variable.}

\item{n_obs}{integer vector. Number of positive observations to make, i.e.
occupancy of simulated species. Can be length 1 or n (default: 1).}

\item{up}{numeric vector. up=uniform proportion. This is the proportion of
the probability distribution P(species) that is composed of a uniform
distribution, if desired. If set to a value above zero (and blow 1), 
P(species) will be a weighted sum of the normal distribution described above,
and a uniform distribution. The weight for the uniform distribution will be
up, and the weight for the normal distribution will be 1-up (default: 0).}

\item{oceanp}{numeric vector. oceanp=ocean proportion. This is the proportion of 
samples in env that are "in the ocean", i.e. samples where the species would
not expect to be found even if env is permissive. If aliens were calculating
specificity of cows to temperature, they might look in the ocean at sites where
the temperature is 17C (great for cows). But cows are not found in the ocean.
This proportion is used to randomly select ocean sites within env, and then
p(s|env|ocean) = up. Can be length 1 or n (default: 0).}

\item{n_cores}{integer. Number of CPU cores for parallel computation (DEFAULT: 2).}

\item{seed}{integer. Seed for randomization. Daughter seeds will be generated for
parallel computations, each with the same number of digits as seed 
(DEFAULT: 1234567).}
}
\value{
List object containing "matrix" and "params" objects:
  \describe{
    \item{matrix:}{
      matrix where each column is a vector of simulated observation frequencies
      (counts) corresponding to a value of env; each row represents a simulated 
      species.
    }
    \item{params:}{
      data.frame of parameters (columns) used to simulate each species (rows).
   }
  }
}
\description{
Simulates inputs for phy_or_env_spec, by creating a species distribution over
an artificial (or real) environmental variable. That distribution has a mean
at the "ideal" environmental value for the simulated species, and the standard
deviation of that distribution controls the extent to which the species is
specific to the variable. A high SD means less specificity, and a low SD means
more specificity.
}
\details{
Since this process can result in failures (if a species is requested that's
highly specific to a region of env that isn't samples), some output species
will be failures. Default operation is to remove those failures from output
matrix and output params data frame, but this can be changed.
}
\examples{
  none yet written.

}
\author{
John L. Darcy
}
