% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_funs.r
\name{fit_gamma_fwd_rev}
\alias{fit_gamma_fwd_rev}
\title{fit_gamma_fwd_rev}
\usage{
fit_gamma_fwd_rev(x, lower_prop = 0.1, upper_prop = 10)
}
\arguments{
\item{x}{numeric vector.}

\item{lower_prop}{float. proportion of gamma max likelihood estimator
to use as lower boundary for fitting (default: 0.10).}

\item{upper_prop}{float. proportion of gamma max likelihood estimator
to use as upper boundary for fitting (default: 10.0).}
}
\value{
a list with par, log likelihood, and type - forward or reverse.
  for reverse, also returns mx which is max(x), needed for reversing
  quantiles.
}
\description{
Fits reversible gamma distribution to x where x is a vector of data.
Internal function for phy_or_env_spec() or calculate_pec_and_pval().
}
\author{
John L. Darcy
}
\keyword{internal}
