% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_fit.r
\name{gamma_fit}
\alias{gamma_fit}
\title{gamma_fit}
\usage{
gamma_fit(x, lower_prop = 0.1, upper_prop = 10)
}
\arguments{
\item{x}{numeric vector of values}

\item{lower_prop}{proportion of starting values to use for the lower bound
of gamma distribution parameters (DEFAULT=0.10). Starting values are
approximated as:
\describe{
  \item{"shape":}{
    (m^2)/v
  }
  \item{"rate":}{
    m/v
  }
}
where m is the mean of x and v is the population variance of x. These values
are multiplied by lower_prop in order to set shape and rate minnima for
constrained optimization. The actual minima are c(0,0), and setting 
lower_prop=0 will acheive this. But since the approximation is pretty
good, using a lower_prop higher than 0 will give a faster optimization.}

\item{upper_prop}{same as lower_prop, but for upper bound (Default = 10).}
}
\value{
Vector of estimated values; c(shape, rate).
}
\description{
Fits a gamma distribution to data. This function is really just a shorter, 
faster, far less felxible version of fitdist() from the "fitdistrplus"
package.
}
\examples{
none yet written.

}
\author{
John L. Darcy
}
