% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_funs.r
\name{gamma_hist_cor}
\alias{gamma_hist_cor}
\title{fit_gamma_fwd_rev}
\usage{
gamma_hist_cor(x, fit)
}
\arguments{
\item{x}{numeric vector.}

\item{fit}{object returned by fit_gamma_fwd_rev().}
}
\value{
a single correlation coefficient.
}
\description{
Calculates user-interpretable check of goodness of fit for gamma
  distribution where x is data, and fit is an object returned by
  fit_gamma_fwd_rev().
}
\author{
John L. Darcy
}
\keyword{internal}
