% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_mode.r
\name{gamma_mode}
\alias{gamma_mode}
\title{gamma_mode}
\usage{
gamma_mode(x, mincor = 0.85, fallback = mean)
}
\arguments{
\item{x}{numeric vector of values}

\item{mincor}{minnimum acceptable correlation between binned frequencies of
x (as in hist(x)) and the predicted probability densities for those bins
from the fit gamma distribution.}

\item{fallback}{function to use on x if gamma fails to fit. Default is mean.
A warning will be given if fallback is used.}
}
\value{
mode of x, single value.
}
\description{
Estimates mode by fitting a gamma distribution.
}
\examples{
  set.seed(12345)
  # example that should work:
  a <- (rgamma(100, 12, 1))
  hist(a)
  gamma_mode(a)
  abline(v=gamma_mode(a), col="red")
  # example that should NOT work:
  a <- runif(100, 0, 1)
  gamma_mode(a)
  

}
\author{
John L. Darcy
}
