% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ga_defaults.r
\name{get_ga_defaults}
\alias{get_ga_defaults}
\title{get_ga_defaults}
\usage{
get_ga_defaults()
}
\value{
named list of parameters.
}
\description{
Simply returns default parameters for the genetic algorithm in rao_genetic_max().
This function has no parameters.
}
\examples{
# get_ga_defaults()

}
\author{
John L. Darcy
}
