% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_funs.r
\name{mode_gamma_fwd_rev}
\alias{mode_gamma_fwd_rev}
\title{mode_gamma_fwd_rev}
\usage{
mode_gamma_fwd_rev(fit, fallback)
}
\arguments{
\item{fit}{object returned by fit_gamma_fwd_rev().}

\item{fallback}{anything. what to return if fit failed.}
}
\value{
either a single value (a mode), or fallback.
}
\description{
calculates mode for a given gamma distribution fit, as produced by
fit_gamma_fwd_rev(). If that function failed, fallback is what to
return if that function instead (as a passthrough).
}
\author{
John L. Darcy
}
\keyword{internal}
