% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_threshold.r
\name{occ_threshold}
\alias{occ_threshold}
\title{occ_threshold}
\usage{
occ_threshold(m, threshold, max_absent = 0)
}
\arguments{
\item{m}{matrix or data frame of numeric values. Columns represent 
species, rows are samples.}

\item{threshold}{integer. Minimum number of samples a species can occupy
without being removed.}

\item{max_absent}{float. Maximum abundance value at which a species will be
considered absent (default: 0).}
}
\value{
matrix with low-occupancy species removed.
}
\description{
removes species (columns) from a matrix that don't meet a minimum occupancy, 
defined as the number of samples in which that species was observed.
}
\examples{
# library(specificity)
# attach(endophyte)
# dim(otutable)
# otutable_over25 <- occ_threshold(otutable, 25)
# dim(otutable_over25)

}
\author{
John L. Darcy
}
