% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_funs.r
\name{p_from_perms_or_gfit}
\alias{p_from_perms_or_gfit}
\title{p_from_perms_or_gfit}
\usage{
p_from_perms_or_gfit(
  emp,
  perm = NULL,
  gfit = NULL,
  fallback = 1,
  tails = 1,
  threshold = 30
)
}
\arguments{
\item{emp}{an empirical value (a quantile).}

\item{perm}{numeric vector of values (use this OR gfit, not both. default: NULL).}

\item{gfit}{object returned by fit_gamma_fwd_rev() (use this OR perm, not both. default: NULL).}

\item{fallback}{anything. what to return if gfit failed (default: 1)}

\item{tails}{integer, 1, 2, or 3. 1 means left tail, 2 means both tails, 3 means right (default: 1).}

\item{threshold}{integer. minimum number of perms required to return a P-value (default: 30).}
}
\value{
either a single value (a P-value), or fallback.
}
\description{
calculates a P-value either from permutations (via the regular quantile approach)
or from a fit_gamma_fwd_rev() fit.
}
\author{
John L. Darcy
}
\keyword{internal}
