% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_funs.r
\name{p_gamma_fwd_rev}
\alias{p_gamma_fwd_rev}
\title{p_gamma_fwd_rev}
\usage{
p_gamma_fwd_rev(q, fit, fallback)
}
\arguments{
\item{q}{a quantile (in non-reversed orientation).}

\item{fit}{object returned by fit_gamma_fwd_rev().}

\item{fallback}{anything. what to return if fit failed.}
}
\value{
either a single value (a P-value), or fallback.
}
\description{
same as pgamma() but for custom fit objects returned by fit_gamma_fwd_rev.
}
\author{
John L. Darcy
}
\keyword{internal}
