% NOT generated by roxygen2, go ahead and edit by hand
\name{pairwise_geo_mean}
\alias{pairwise_geo_mean}
\title{pairwise_geo_mean}
\usage{
pairwise_geo_mean(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
vector of pairwise geometric means, of length (l*l-l)/2, where l=length(x).
}
\description{
Calculates pairwise geometric means from unique 2-element combinations of 
vector x. Written in C++ because R slow. The output vector is the same length 
and same order as a lower triangle of matrix with rows and columns x.
}
\examples{
  x <- 1:6
  y_cpp <- pairwise_geo_mean(x)
  y_r <- as.dist(outer(x, x, function(x,y){sqrt(x*y)}))
  print("Calculated with R's outer() function:")
  y_r
  print("As a vector:")
  as.vector(y_r)
  print("Calculated with pairwise_geo_mean (C++):")
  y_cpp
}
\author{
John L. Darcy
}