% NOT generated by roxygen2, go ahead and edit by hand
\name{pairwise_product}
\alias{pairwise_product}
\title{pairwise_product}
\usage{
pairwise_product(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
vector of pairwise_products, of length (l^2-l)/2, where l=length(x).
}
\description{
Calculates pairwise_products from unique 2-element combinations of 
vector x. The output vector is the same length and same order as a
lower triangle of matrix with rows and columns x.
}
\examples{
  # x <- 1:6
  # y_cpp <- pairwise_product(x)
  # y_r <- as.dist(outer(x, x, function(x,y){x*y}))
  # print("Calculated with R's outer() function:")
  # y_r
  # print("As a vector:")
  # as.vector(y_r)
  # print("Calculated with pairwise_product (C++):")
  # y_cpp
}
\author{
John L. Darcy
}