% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30__phy_or_env_spec.r
\name{phy_or_env_spec}
\alias{phy_or_env_spec}
\title{phy_or_env_spec}
\usage{
phy_or_env_spec(abunds_mat, env = NULL, hosts = NULL,
  hosts_phylo = NULL, n_sim = 1000, sim_fun = function(m) {    
  m[sample(1:nrow(m)), ] }, p_adj = "fdr", seed = 1234567, tails = 1,
  n_cores = 2, verbose = TRUE, lowmem = FALSE, vec_wvar = FALSE,
  phy_ent = FALSE, denom_type = "global_unif")
}
\arguments{
\item{abunds_mat}{matrix or data frame of numeric values. Columns represent 
species, rows are samples. For columns where the value is nonzero for two or
fewer data points, environmental SES cannot be calculated, and NAs will be 
returned. Negative values in abunds_mat are not allowed (REQUIRED).}

\item{env}{numeric vector, dist, or square matrix. Environmental variable 
corresponding to abunds. For example, temperature, or geographic distance.
Not required for computing phylogenetic specificity (DEFAULT: NULL).}

\item{hosts}{character vector. Host identities corresponding to abunds. Only
required if calculating SES for phylogenetic specificity (DEFAULT: NULL).}

\item{hosts_phylo}{phylo object. Tree containing all unique hosts as tips. Only
required if calculating SES for phylogenetic specificity (DEFAULT: NULL).}

\item{n_sim}{integer. Number of simulations of abunds_mat to do under the null 
hypothesis that host or environmental association is random. P-values will not
be calculated if n_sim < 100 (DEFAULT: 500).}

\item{sim_fun}{function. A function f where f(abunds_mat) returns a matrix
object with the same number of rows and columns as abunds_mat. Default is
f=function(m){ m[sample(1:nrow(m)),]}, which just permutes the order of 
rows in abunds_mat. Users may wish to use a null model that is able to
preserve row and column totals such as the function permatswap() from the
vegan package or the function vaznull() from the bipartite package. Either
of these can be easily adapted to return only a single matrix (see examples). 
However, neither can accomodate non-integer matrices.}

\item{p_adj}{string. Type of multiple hypothesis testing correction performed
on P-values. Can take any valid method argument to p.adjust, including "none",
"bonferroni", "holm", "fdr", and others (DEFAULT: "fdr").}

\item{seed}{integer. Seed to use so that this is repeatable (DEFAULT: 1234557).}

\item{tails}{integer. 1 = 1-tailed, test for specificity only. 2 = 2-tailed.
3 = 1-tailed, test for cosmopolitanism only. 0 = no test, P=1.0 (DEFAULT: 1).}

\item{n_cores}{integer. Number of CPU cores to use for parallel operations. If
set to 1, lapply will be used instead of mclapply (DEFAULT: 2).}

\item{verbose}{logical. Should status messages be displayed? (DEFAULT: TRUE).}

\item{lowmem}{logical. Should this function be run in a way that saves memory?
If TRUE, will run FAR slower but use almost no ram. If FALSE, lots of ram is
used but it will run quickly. For example, on an 800x1600 matrix, with nperm=
1000, n_cores=12 and lowmem=F, will use roughly 90 GB ram but finish in one
minute. With the same settings but lowmem=T, will use less than 1 GB ram but 
take over an hour. A progress bar is shown if TRUE (DEFAULT: FALSE).}

\item{vec_wvar}{logical. For vector input, should old weighted-variance approach
be used instead of transforming the vector into a euclidean distance matrix?
(DEFAULT: FALSE).}

\item{phy_ent}{logical. For phylogenetic input, should SES of phylogenetic
entropy be used instead of trandforming the phylogeny into patristic distances
using tree2mat? (DEFAULT: FALSE).}

\item{denom_type}{string. Type of denominator (d) to use for SES. SES is the 
difference between empirical and mean simulated specificity, divided by d
(DEFAULT: global_unif).
\describe{
  \item{species_sim:}{
    d for species s is calculated as the standard deviation of specificities
    calculated from permuted abundances of s. This makes SES of specificity
    counterintuitively sensitive to occupancy, where species with high occupancy
    have more extreme SES of specificity than rare species, due to their more 
    deterministic sim specificities. Not suggested.
  }
  \item{global_sim:}{
    d is same for all species. Calculated as standard deviation of all
    specificities calculated from permuted species. SES is much less sensitive
    to occupancy, and intuitively, species with greater occupancy have less
    extreme SES of specificity than rare species. However, using this d makes
    results from different abundance matrices NOT comparable, because d is a
    function of the distribution of species distributions unique to abunds_mat.
    Results for different variables and the same abundance matrix are still 
    comparable, though (obviously using same denom_type and other parameters).
  }
  \item{global_unif:}{
    d is same for all species. Calculated as variability in specificity under
    random uniform distribution (beta 1,1) of species abunds. This d is 
    comparable between different abundance matrices and between different
    variables. SES results using this denom_type are NOT comparable with results
    from any other. Due to its stability, this is the default option.
  }
  \item{raw:}{
    d is 1 for all species, so SES is not actually standardized (although the
    column label in output data will still be "SES"). This option only exists
    for diagnostic purposes, since units of this metric are incomprehensible.
  }
}}
}
\value{
data.frame where each row is an input species. First column is P-value
  ($Pval), second column is SES ($SES).
}
\description{
Calculates standardized effect size (SES) of species' specificity to either
a 1-dimensional variable (vector), 2-dimensional variable (matrix), or to
a phylogeny. Default operation is to transform all variable input types into a 
matrix M, and calculate specificity as SES of the empirical sum of M
weighted by weights matrix W. W is a matrix of pair-wise products of species
species abundances (Rao's quadratic entropy). Optionally, host phylogenetic
specificity can be calculated per Poulin et al. (2011) using weighted 
phylogenetic entropy, and specificity to 1-dimensional data can be calculated
using a weighted variance approach. These approaches are included mainly for
comparative methodological purposes.
}
\examples{
# phylogenetic specificity using endophyte data set
attach(endophyte)
# only analyze species with occupancy >= 20
m <- occ_threshold(prop_abund(zotutable), 20)
ses_host <- phy_or_env_spec(
    abunds_mat=m,
    hosts=metadata$PlantGenus, 
    hosts_phylo=supertree,
    n_cores=12
)

# using vazquez null model from bipartite package as an alternate permutation:
# note that the "creating permuted matrices" step will be slow.
library(bipartite)
ses_host_vaz <- phy_or_env_spec(
    abunds_mat=m,
    hosts=metadata$PlantGenus, 
    hosts_phylo=supertree,
    n_cores=12,
    sim_fun=function(m){bipartite::vaznull(1, m)[[1]]},
)

# compare naive permutation vs. vazquez:
plot(ses_host$SES, ses_host_vaz$SES, ylab="bipartite::vaznull", xlab="naive")
abline(h=0);abline(v=0)
hist(ses_host_vaz$SES)
hist(ses_host$SES)

# environmental specificity using elevation from endophyte data set:
ses_elev <- phy_or_env_spec(
    abunds_mat=m,
    env=metadata$Elevation,
    n_cores=12
)

# geographic specificity using spatial data from endophyte data set:
ses_geo <- phy_or_env_spec(
    abunds_mat=m,
    env=distcalc(metadata$Lat, metadata$Lon),
    n_cores=12
)

}
\references{
Poulin et al. (2011) Host specificity in phylogenetic and geographic
  space. Trends Parasitol 8:355-361. doi: 10.1016/j.pt.2011.05.003
}
\author{
John L. Darcy
}
