% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_spec_sim.r
\name{phy_spec_sim}
\alias{phy_spec_sim}
\title{phy_spec_sim}
\usage{
phy_spec_sim(
  sdev,
  ideal,
  ideal2 = "",
  ideal3 = "",
  n_ideal = 1,
  hosts,
  hosts_phylo,
  n_obs,
  up = 0,
  oceanp = 0,
  n_cores = 2,
  seed = 1234567
)
}
\arguments{
\item{sdev}{numeric vector. Standard deviation of the probability distribution
P(s), in units of patristic distance in hosts_phylo. Low values  mean that
species s is found with a narrow grouping of hosts, i.e. specificity. High
values mean that s is found across a wider group of hosts, i.e.
cosmopolitanism. Multiple values can be input in order to simulate a range
of specificities, simultaneously. To get a handle on this somewhat opaque 
variable, consider plotting a histogram of patristic distances within 
hosts_phylo (see: ape::cophenetic.phylo). Can be length 1 or n.}

\item{ideal}{character vector. Tip label of hosts_phylo that is ideal (or
closest to ideal) for the simulated species. Does not have to be in hosts,
but MUST be in hosts_phylo. Can be length 1 or n.}

\item{ideal2}{character vector. Tip label of hosts_phylo that is secondary
ideal host for the simulated species. Does not have to be in hosts, but MUST
be in hosts_phylo. Can be blank ("") if corresponding n_ideal < 2. Can be
length 1 or n (default: "").}

\item{ideal3}{character vector. Tip label of hosts_phylo that is tertiary
ideal host for the simulated species. Does not have to be in hosts, but MUST
be in hosts_phylo. Can be blank ("") if corresponding n_ideal < 3. Can be
length 1 or n (default: "").}

\item{n_ideal}{integer vector. number of ideal hosts to use. Must be 1, 2, or
3 (default: 1).}

\item{hosts}{character vector. Real of fake host identities. All must be 
tips within hosts_phylo. Analogous to env argument to env_spec_sim.}

\item{hosts_phylo}{phylo object. Tree containing all unique hosts as tips.}

\item{n_obs}{integer vector. Number of positive observations to make, i.e.
occupancy of simulated species. Can be length 1 or n.}

\item{up}{numeric vector. up=uniform proportion. This is the proportion of
the probability distribution P(species) that is composed of a uniform
distribution, if desired. If set to a value above zero (and blow 1), 
P(species) will be a weighted sum of the normal distribution described above,
and a uniform distribution. The weight for the uniform distribution will be
up, and the weight for the normal distribution will be 1-up (default: 0).}

\item{oceanp}{numeric vector. See ?env_spec_sim for help.}

\item{n_cores}{integer. Number of CPU cores for parallel computation (default: 2).}

\item{seed}{integer. Seed for randomization. Daughter seeds will be generated for
parallel computations, each with the same number of digits as seed 
(default: 1234567).}
}
\value{
List object containing "matrix" and "params" objects:
  \describe{
    \item{matrix:}{
      matrix where each column is a vector of simulated observations corresponding
      to a value of hosts; each row represents a simulated species.
    }
    \item{params:}{
      data.frame of parameters (columns) used to simulate each species (rows).
      A column called "index" is included so that simulated species can be mapped
      back onto original data structures when some species are ommitted due to 
      simulation failure (see fail_rm).
    }
  }
}
\description{
Simulates inputs for phy_or_env_spec, by creating a species distribution over
an artificial (or real) host phylogenetic tree. For a phylogeny, the species
probability distribution P(s) is based on patristic distances within the
tree, such that P(s) is maximized at zero patristic distance between a tip
in the tree and the ideal host species for s. This distribution is given by a
a truncated normal distribution centered on zero, using only positive values.
A uniform proportion (up) to that distribution may be added as well, to add a
baseline probability to P(s). The standard deviation of P(s) can be raised or
lowered to simulate cosmopolitanism or specificity.
}
\examples{
  none yet written.

}
\author{
John L. Darcy
}
