% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid_abunds.r
\name{plot_grid_abunds}
\alias{plot_grid_abunds}
\title{plot_grid_abunds}
\usage{
plot_grid_abunds(grid, abunds, pch = "", ...)
}
\arguments{
\item{grid}{data frame with columns x and y, representing cartesian coordinates
of sample locations. Can be artificial (generate with randomgrid()) or real.}

\item{abunds}{abundances of a species, corresponding to rows in grid.}

\item{pch}{pch character code to use for bottom of each abundance line (default: "")}

\item{...}{arguments to be passed to plot.}
}
\value{
returns nothing, just makes a plot.
}
\description{
plots species abundances across spatial sampling locations
}
\examples{
library(specificity)
g1 <- randomgrid()
plot(g1)
a1 <- geo_spec_sim(sdev=c(30, 30, 30, 30), n_obs=1000, 
  grid=g1, up=c(0, 0.20, 0.40, 0.60))
par(mfrow=c(2,2))
plot_grid_abunds(g1, a1$matrix[,1])
plot_grid_abunds(g1, a1$matrix[,2])
plot_grid_abunds(g1, a1$matrix[,3])
plot_grid_abunds(g1, a1$matrix[,4])

}
\author{
John L. Darcy
}
