% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specificities.r
\name{plot_specificities}
\alias{plot_specificities}
\title{plot_specificities}
\usage{
plot_specificities(specs_list, n_bins = 20, col_sig = "black",
  col_nsig = "gray", col_bord = NA, alpha = 0.05, label_cex = 0.6)
}
\arguments{
\item{specs_list}{list of data.frames. Each data.frame must be an output from
phy_or_env_spec; must have columns "Spec" and "Pval".}

\item{n_bins}{integer. Number of bins for stacked violins (DEFAULT: 20).}

\item{col_sig}{string. Color name or hex code for species where Pval <= alpha
(DEFAULT = "black").}

\item{col_nsig}{string. Color name or hex code for species where Pval > alpha
(DEFAULT = "gray").}

\item{col_bord}{string. Color name or hex code for border color. Use NA for no
border (DEFAULT = NA).}

\item{alpha}{float. alpha value for determining statistical significance; see
col_sig and col_nsig above (DEFAULT = 0.05).}
}
\value{
returns nothing (a plot is made).
}
\description{
Visualizes results from phy_or_env_spec
}
\examples{
  none yet written.

}
\author{
John L. Darcy
}
