% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30__phy_or_env_spec.r
\name{pval_from_perms}
\alias{pval_from_perms}
\title{pval_from_perms}
\usage{
pval_from_perms(emp, perm, tails, threshold = 100)
}
\arguments{
\item{emp}{Numeric scalar. An empirical test statistic value.}

\item{perm}{Numeric vector. Test statistic values similar to emp, but calculated
from permuted data.}

\item{tails}{integer. 
\describe{
  \item{1:}{Left tail only.}
  \item{2:}{2-tailed test.}
  \item{3:}{Right tail only.}
  \item{0:}{No test, P=1.}
}}

\item{threshold}{integer. Minimum number n of non-NA values in perm that are
acceptable. If n < threshold, P=NA (DEFAULT: 100).}
}
\value{
a P-value.
}
\description{
Calculates P-value for permutation tests.
}
\author{
John L. Darcy
}
