% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomgrid.r
\name{randomgrid}
\alias{randomgrid}
\title{randomgrid}
\usage{
randomgrid(
  n_samp = 1000,
  xmin = -100,
  xmax = 100,
  ymin = -100,
  ymax = 100,
  seed = 123456
)
}
\arguments{
\item{n_samp}{number of sampling locations to output (default: 1000).}

\item{xmin}{minimum x-axis coordinate (default: -100).}

\item{xmax}{maximum x-axis coordinate (default: 100).}

\item{ymin}{minimum y-axis coordinate (default: -100).}

\item{ymax}{maximum y-axis coordinate (default: 100).}

\item{seed}{integer, seed for randomization.}
}
\value{
data.frame object with x and y columns, with n_samp rows.
}
\description{
Generates a random spatial sampling using a bivariate random uniform distribution.
}
\examples{
library(specificity)
g <- randomgrid()
plot(g)
g2 <- randomgrid(n_samp=50, xmin=0, ymin=0)
plot(g2)

}
\author{
John L. Darcy
}
