% NOT generated by roxygen2, go ahead and edit by hand
\name{rao1sp}
\alias{rao1sp}
\title{rao1sp}
\usage{
rao1sp(
	p, 
	D, 
	perm=FALSE, 
	seed=0)
}
\arguments{
\item{p}{numeric vector of length n. A species weights vector.}
\item{D}{numeric vector of length n(n-1)/2. i.e. a dist object representing the lower triangle of an nxn matrix.}
\item{perm}{bool. Whether or not the permute order of p before calculating Rao (default: FALSE).}
\item{seed}{integer. a seed to be used if perm=TRUE. setting seed=0 will give nondeterministic random results, as if no seed were set (default: 0).}
}
\value{
A single Rao value.
}\description{
Calculate's Rao's quadratic entropy for one species (rao1sp = "rao one species"). 
}
\examples{
  # library(specificity)
  # p <- 1:5/sum(1:5)
  # D <- dist(6:10)
  # rao1sp(p, D)
}
\author{
John L. Darcy
}
