% NOT generated by roxygen2, go ahead and edit by hand
\name{rao_genetic_max}
\alias{rao_genetic_max}
\title{rao_genetic_max}
\usage{
rao_genetic_max(p, D)
}
\arguments{
\item{p}{numeric vector of length n - a species weights vector.}
\item{D}{numeric vector of length n(n-1)/2 - i.e. a dist object whose full matrix is nxn.}
\item{term_cycles}{integer, number of cycles with no improvement to trigger termination (DEFAULT=10).}
\item{maxiters}{integer, maximum number of iterations to run algorithm (DEFAULT=400).}
\item{popsize}{integer, population size for genetic algorithm (DEFAULT=300).}
\item{keep}{integer, number of individuals to keep during each iteration (DEFAULT=5).}
\item{prc}{double, precision for calculating termination with term_cycles (DEFAULT=0.001).}
}
\value{
List object containing results of genetic algorithm:
  \describe{
    \item{best_rao:}{
      Maximum Rao value found.
    }
    \item{iter_raos:}{
      Max Rao value for each iteration. If termination condition was met, rest of values after final iteration are NA.
    }
    \item{iterations:}{
      Iteration numbers, corresponding to iter_raos.
    }
    \item{best_p:}{
      The best permutation of p found (corresponds to best_rao).
    }
  }
}\description{
Uses a genetic algorithm to find the optimum permutation of p to maximize Rao(p,D). 
}
\examples{
  set.seed(12345)
  p <- runif(100) * 20
  D <- dist(sample(p))
  a <- rao_genetic_max(p,D)
  plot(a$iter_raos ~ a$iterations)
}
\author{
John L. Darcy
}
