% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10__wpd.r
\name{rao_quad_ent}
\alias{rao_quad_ent}
\title{rao_quad_ent}
\usage{
rao_quad_ent(d, w, raw = FALSE)
}
\arguments{
\item{d}{numeric dist. Distances, as a dist object. Note that dist objects can easily
be made from square matrices using as.dist(), or euclidean distances can be calculated
from numeric vectors using dist().}

\item{w}{numeric vector. Per-sample weights, as a vector. For example, abundances of
a species across samples. w MUST be sorted such that it corresponds to rows of
as.matrix(d). Thus, w must have length l such that (l^2-l)/2 = length(d).}

\item{raw}{logical. If true, FDq will be returned. If false, FDq* will be returned
(DEFAULT: FALSE).}
}
\value{
A single value.
}
\description{
Calculates Rao's (1982) quadratic entropy (FDq) from a distance matrix and a vector of
weights (e.g. relative abundance data). In simple terms, FDq is sum product of distances
and pairwise products of weights. Default operation in this function is to then divide 
FDq by the sum of pairwise weights, to give a weighted mean of distances (FDq*). This 
gives units of distance, and also makes the metric insensitive to the sum of weights, 
i.e. weights do not need to be normalized before calculation. This behavior can be
disabled by setting raw=TRUE, which will give FDq.
}
\examples{
  none yet written

}
\references{
Rao R (1982) Diversity: its measurement, decomposition, apportionment and
  analysis. Sankhyā: The Indian Journal of Statistics 44(1).
}
\author{
John L. Darcy
}
