% NOT generated by roxygen2, go ahead and edit by hand
\name{rao_sort_max}
\alias{rao_sort_max}
\title{rao_sort_max}
\usage{
rao_sort_max(w, D)
}
\arguments{
\item{w}{numeric vector of length n - a species weights vector.}
\item{D}{numeric vector of length n(n-1)/2 - i.e. a dist object whose full matrix is nxn.}
}
\value{
A single value, approximating maximum rao under permutations of w.
}\description{
Sorts pairwise_product(w) and D to approximate the maximum of Rao(w,D) under permutations of w.
}
\examples{
  # none yet written.
}
\author{
John L. Darcy
}
