% NOT generated by roxygen2, go ahead and edit by hand
\name{raoperms}
\alias{raoperms}
\title{raoperms}
\usage{
raoperms(
	p,
	D,
	n_sim=1000,
	seed=12345)
}
\arguments{
\item{p}{numeric vector of length n. A species weights vector.}
\item{D}{numeric vector of length n(n-1)/2. i.e. a dist object representing the lower triangle of an nxn matrix.}
\item{n_sim}{integer. number of sims to do (default: 1000).}
\item{seed}{integer. a seed to be used for permutation. setting seed=0 will give nondeterministic random results, as if no seed were set (default: 0).}
}
\value{
Vector of Rao values, length = n_sim.
}
\description{
C++ function used by phy_or_env_spec(). Not meant for use otherwise.
}
\examples{
  # library(specificity)
  # p <- runif(100) 
  # D <- dist(runif(100))
  # a <- raoperms(p,D,100,12345)
  # hist(a)
}

\author{
John L. Darcy
}