% NOT generated by roxygen2, go ahead and edit by hand
\name{spec_core}
\alias{spec_core}
\title{spec_core}
\usage{
spec_core(w,D)
}
\arguments{
\item{w}{numeric matrix with m rows and n columns.}
\item{D}{numeric vector of length (n(n-1))/2.}
}
\value{
Vector of Rao specificities corresponding to columns of w.
}
\description{
C++ function used by phy_or_env_spec(). Not meant for use otherwise.
}
\author{
John L. Darcy
}