\name{dum, eta2, scramble, specEta2, sepcLm, specGlm, print.summary.specificity, print.summary.specificity}
\alias{dum}
\alias{eta2}
\alias{scramble}
\alias{specEta2}
\alias{specGlm}
\alias{specLm}
\alias{print.summary.specificity}
\alias{summary.specificity}
\title{
Helper functions for \code{specificity}
}
\description{
Helpers for \code{specificity: dum, eta2, scramble, specEta2, specGlm, specLm, print.specificitySummary, summary.specificity}.
}
