% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30__phy_or_env_spec.r
\name{check_pes_inputs}
\alias{check_pes_inputs}
\title{check_pes_inputs}
\usage{
check_pes_inputs(abunds_mat, env, hosts, hosts_phylo, verbose = TRUE)
}
\arguments{
\item{abunds_mat}{(required, see phy_or_env_spec)}

\item{env}{(required, can be NULL, see phy_or_env_spec)}

\item{hosts}{(required, can be NULL, see phy_or_env_spec)}

\item{hosts_phylo}{(required, can be NULL, see phy_or_env_spec)}

\item{verbose}{logical. Should status messages be displayed? (DEFAULT: TRUE).}
}
\value{
string. either "mat", "dist", "vec", "phy", or "error".
}
\description{
Function used by phy_or_env_spec. 
checks abunds_mat, env, hosts, and hosts_phylo inputs to phy_or_env_spec to
make sure there are no problems. This could include missing species in trees,
incompatible dimensions, non-numeric inputs, etc. Returns an input type, which
is just a string that can be "mat", "dist", "vec", "phy", or "error".
}
