% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30__phy_or_env_spec.r
\name{phy_or_env_spec}
\alias{phy_or_env_spec}
\title{phy_or_env_spec}
\usage{
phy_or_env_spec(abunds_mat, env = NULL, hosts = NULL,
  hosts_phylo = NULL, n_sim = 1000, sim_fun = function(m) {    
  m[sample(1:nrow(m)), ] }, p_adj = "fdr", seed = 1234567, tails = 1,
  n_cores = 2, verbose = TRUE, lowmem = FALSE, p_method = "raw",
  denom_type = "index", matrix_tmax = 1, diagnostic = F)
}
\arguments{
\item{abunds_mat}{matrix or data frame of numeric values. Columns represent 
species, rows are samples. For columns where the value is nonzero for two or
fewer data points, environmental SES cannot be calculated, and NAs will be 
returned. Negative values in abunds_mat are not allowed (REQUIRED).}

\item{env}{numeric vector, dist, or square matrix. Environmental variable 
corresponding to abunds. For example, temperature, or geographic distance.
Not required for computing phylogenetic specificity (DEFAULT: NULL).}

\item{hosts}{character vector. Host identities corresponding to abunds. Only
required if calculating SES for phylogenetic specificity (DEFAULT: NULL).}

\item{hosts_phylo}{phylo object. Tree containing all unique hosts as tips. Only
required if calculating SES for phylogenetic specificity (DEFAULT: NULL).}

\item{n_sim}{integer. Number of simulations of abunds_mat to do under the null 
hypothesis that host or environmental association is random. P-values will not
be calculated if n_sim < 100 (DEFAULT: 500).}

\item{sim_fun}{function. A function f where f(abunds_mat) returns a matrix
object with the same number of rows and columns as abunds_mat. Default is
f=function(m){ m[sample(1:nrow(m)),]}, which just permutes the order of 
rows in abunds_mat. Users may wish to use a null model that is able to
preserve row and column totals such as the function permatswap() from the
vegan package or the function vaznull() from the bipartite package. Either
of these can be easily adapted to return only a single matrix (see examples). 
However, neither can accomodate non-integer matrices.}

\item{p_adj}{string. Type of multiple hypothesis testing correction performed
on P-values. Can take any valid method argument to p.adjust, including "none",
"bonferroni", "holm", "fdr", and others (DEFAULT: "fdr").}

\item{seed}{integer. Seed to use so that this is repeatable (DEFAULT: 1234557).}

\item{tails}{integer. 1 = 1-tailed, test for specificity only. 2 = 2-tailed.
3 = 1-tailed, test for cosmopolitanism only. 0 = no test, P=1.0 (DEFAULT: 1).}

\item{n_cores}{integer. Number of CPU cores to use for parallel operations. If
set to 1, lapply will be used instead of mclapply (DEFAULT: 2).}

\item{verbose}{logical. Should status messages be displayed? (DEFAULT: TRUE).}

\item{p_method}{string. method argument to pval_from_perms (DEFAULT: "raw").}

\item{matrix_tmax}{float. Theoretical maximum value of distances when a matrix or dist
is supplied as env. Only used with denom_type = "index" (the default); if so an error
will be thrown if there is a value inside a supplied matrix greater than matrix_tmax
(DEFAULT: 1).}

\item{diagnostic}{logical. If true, changes output to include different parts of SES. 
This includes Pval, SES, raw, denom, emp, and all sim values with column labels as
simN where N is the number of sims (DEFAULT: FALSE)}
}
\value{
data.frame where each row is an input species. First column is P-value
  ($Pval), second column is specificity ($Spec).
}
\description{
Calculates species' specificities to either a 1-dimensional variable (vector), 
2-dimensional variable (matrix), or to a phylogeny. Transforms all variable
input types into a matrix D, and calculates specificity by comparing empirical
RQE* (weighted mean of D's lower triangle and unique pairwise products of 
species abundances) to simulated RQE* (same but with permuted abundances). This
"raw" specificity is then divided by some denominator d in order to standardize
it such that specificities can be compared between different species and
different variables. Values closer to 0 indicate random assortment (null
hypothesis), and more negative values indicate stronger specificity.
}
\examples{
# phylogenetic specificity using endophyte data set
attach(endophyte)
# only analyze species with occupancy >= 20
m <- occ_threshold(prop_abund(zotutable), 20)
ses_host <- phy_or_env_spec(
    abunds_mat=m,
    hosts=metadata$PlantGenus, 
    hosts_phylo=supertree,
    n_cores=12
)

# using vazquez null model from bipartite package as an alternate permutation:
# note that the "creating permuted matrices" step will be slow.
library(bipartite)
ses_host_vaz <- phy_or_env_spec(
    abunds_mat=m,
    hosts=metadata$PlantGenus, 
    hosts_phylo=supertree,
    n_cores=12,
    sim_fun=function(m){bipartite::vaznull(1, m)[[1]]},
)

# compare naive permutation vs. vazquez:
plot(ses_host$Spec, ses_host_vaz$Spec, ylab="bipartite::vaznull", xlab="naive")
abline(h=0);abline(v=0)
hist(ses_host_vaz$Spec)
hist(ses_host$Spec)

# environmental specificity using elevation from endophyte data set:
ses_elev <- phy_or_env_spec(
    abunds_mat=m,
    env=metadata$Elevation,
    n_cores=12
)

# geographic specificity using spatial data from endophyte data set:
ses_geo <- phy_or_env_spec(
    abunds_mat=m,
    env=distcalc(metadata$Lat, metadata$Lon),
    n_cores=12
)

}
\references{
Poulin et al. (2011) Host specificity in phylogenetic and geographic
  space. Trends Parasitol 8:355-361. doi: 10.1016/j.pt.2011.05.003
}
\author{
John L. Darcy
}
