% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_or_env_spec.r
\name{phy_or_env_spec}
\alias{phy_or_env_spec}
\title{phy_or_env_spec}
\usage{
phy_or_env_spec(abunds_mat, env = NULL, hosts = NULL,
  hosts_phylo = NULL, n_sim = 1000, sim_fun = function(m) {    
  m[sample(1:nrow(m)), ] }, p_adj = "fdr", seed = 1234567, tails = 1,
  n_cores = 2, verbose = TRUE, p_method = "raw",
  denom_type = "index", diagnostic = F)
}
\arguments{
\item{abunds_mat}{matrix or data frame of numeric values. Columns represent 
species, rows are samples. For columns where the value is nonzero for two or
fewer data points, environmental SES cannot be calculated, and NAs will be 
returned. Negative values in abunds_mat are not allowed (REQUIRED).}

\item{env}{numeric vector, dist, or square matrix. Environmental variable 
corresponding to abunds. For example, temperature, or geographic distance.
Not required for computing phylogenetic specificity (DEFAULT: NULL).}

\item{hosts}{character vector. Host identities corresponding to abunds. Only
required if calculating SES for phylogenetic specificity (DEFAULT: NULL).}

\item{hosts_phylo}{phylo object. Tree containing all unique hosts as tips. Only
required if calculating SES for phylogenetic specificity (DEFAULT: NULL).}

\item{n_sim}{integer. Number of simulations of abunds_mat to do under the null 
hypothesis that host or environmental association is random. P-values will not
be calculated if n_sim < 100 (DEFAULT: 500).}

\item{sim_fun}{function. A function f where f(abunds_mat) returns a matrix
object with the same number of rows and columns as abunds_mat. Default is
f=function(m){ m[sample(1:nrow(m)),]}, which just permutes the order of 
rows in abunds_mat. Users may wish to use a null model that is able to
preserve row and column totals such as the function permatswap() from the
vegan package or the function vaznull() from the bipartite package. Either
of these can be easily adapted to return only a single matrix (see examples). 
However, neither can accomodate non-integer matrices.}

\item{p_adj}{string. Type of multiple hypothesis testing correction performed
on P-values. Can take any valid method argument to p.adjust, including "none",
"bonferroni", "holm", "fdr", and others (DEFAULT: "fdr").}

\item{seed}{integer. Seed to use so that this is repeatable (DEFAULT: 1234557).}

\item{tails}{integer. 1 = 1-tailed, test for specificity only. 2 = 2-tailed.
3 = 1-tailed, test for cosmopolitanism only. 0 = no test, P=1.0 (DEFAULT: 1).}

\item{n_cores}{integer. Number of CPU cores to use for parallel operations. If
set to 1, lapply will be used instead of mclapply (DEFAULT: 2).}

\item{verbose}{logical. Should status messages be displayed? (DEFAULT: TRUE).}

\item{p_method}{string. method argument to pval_from_perms (DEFAULT: "raw").}

\item{denom_type}{string. Type of denominator (d) to use (DEFAULT: "index"). Note
that denominator type does NOT affect P-values.
\describe{
  \item{"ses":}{
    d for species s is calculated as the standard deviation of RQE values
    calculated from permuted species weights. This makes the output specificity
    a standardized effect size (SES). Unfortunately, this makes SES 
    counterintuitively sensitive to occupancy, where species with high occupancy
    have more extreme SES than rare species, due to their more deterministic sim
    specificities. Included for comparative purposes, not suggested.
  }
  \item{"raw":}{
    d is 1 for all species, so output specificity has units of distance, i.e. the
    raw difference between empirical and simulated RQE. This means that results
    from different variables are not comparable, since it is not scale-invariant to
    env or hosts_phylo. It not scale-invariant to the species weights in aunds_mat,
    either. Not sensitive to number of samples. Not suggested because units are
    strange, and isn't comparable between variables. 
  }
  \item{"index":}{
    d is the mean of simulated (permuted) RQE values for species that have stronger
    specificity than expected by chance, resulting in specificity values with range
    [-1, 0), with 0 as the null hypothesis. In this case, -1 indicates perfect
    specificity, where a species is associated with zero environmental variability.
    In the euclidean sense, this could be a species that is always found at the
    exact same elevation or the exact same pH. For species that have weaker specificity
    than expected by chance, d is x minus the mean of simulated RQE values, where x is
    the maximum possible dissimilarity observable given species weights. This d has 
    other useful properties: scale invariance to env/hosts_phylo, insensitivity to 
    the number of samples, insensitivity to occupancy, and strong sensitivity to 
    specificity (DEFAULT).
  }
}}

\item{diagnostic}{logical. If true, changes output to include different parts of SES. 
This includes Pval, SES, raw, denom, emp, and all sim values with column labels as
simN where N is the number of sims (DEFAULT: FALSE)}
}
\value{
data.frame where each row is an input species. First column is P-value
  ($Pval), second column is specificity ($Spec).
}
\description{
Calculates species' specificities to either a 1-dimensional variable (vector), 
2-dimensional variable (matrix), or to a phylogeny. Transforms all variable
input types into a matrix D, and calculates specificity by comparing empirical
Rao's Quadratic Entropy to simulated RQE (same but with permuted abundances). 
By default (denom_type = "index"), an index is calculated from emp and sim
values such that Spec=0 indicates random assortment (null hypothesis), and more
negative values indicate stronger specificity.
}
\examples{
# phylogenetic specificity using endophyte data set
attach(endophyte)
# only analyze species with occupancy >= 20
m <- occ_threshold(prop_abund(zotutable), 20)
ses_host <- phy_or_env_spec(
    abunds_mat=m,
    hosts=metadata$PlantGenus, 
    hosts_phylo=supertree,
    n_cores=12
)

# using vazquez null model from bipartite package as an alternate permutation:
# note that the "creating permuted matrices" step will be slow.
library(bipartite)
ses_host_vaz <- phy_or_env_spec(
    abunds_mat=m,
    hosts=metadata$PlantGenus, 
    hosts_phylo=supertree,
    n_cores=12,
    sim_fun=function(m){bipartite::vaznull(1, m)[[1]]},
)

# compare naive permutation vs. vazquez:
plot(ses_host$Spec, ses_host_vaz$Spec, ylab="bipartite::vaznull", xlab="naive")
abline(h=0);abline(v=0)
hist(ses_host_vaz$Spec)
hist(ses_host$Spec)

# environmental specificity using elevation from endophyte data set:
ses_elev <- phy_or_env_spec(
    abunds_mat=m,
    env=metadata$Elevation,
    n_cores=12
)

# geographic specificity using spatial data from endophyte data set:
ses_geo <- phy_or_env_spec(
    abunds_mat=m,
    env=distcalc(metadata$Lat, metadata$Lon),
    n_cores=12
)

}
\references{
Poulin et al. (2011) Host specificity in phylogenetic and geographic space. 
  Trends Parasitol 8:355-361. doi: 10.1016/j.pt.2011.05.003
Rao CR (2010) Quadratic entropy and analysis of diversity. Sankhya 72:70-80.
  doi: 10.1007/s13171-010-0016-3
Rao CR (1982) Diversity and dissimilarity measurements: A unified approach.
  Theor Popul Biol 21:24-43.
}
\author{
John L. Darcy
}
