% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_or_env_spec.r
\name{phy_or_env_spec}
\alias{phy_or_env_spec}
\title{phy_or_env_spec}
\usage{
phy_or_env_spec(
  abunds_mat,
  env = NULL,
  hosts = NULL,
  hosts_phylo = NULL,
  n_sim = 1000,
  p_adj = "fdr",
  seed = 1234567,
  tails = 1,
  n_cores = 2,
  verbose = TRUE,
  p_method = "raw",
  center = "mean",
  denom_type = "index",
  diagnostic = F
)
}
\arguments{
\item{abunds_mat}{matrix or data frame of numeric values. Columns represent 
species, rows are samples. For columns where the value is nonzero for two or
fewer data points, specificity cannot be calculated, and NAs will be 
returned. Negative values in abunds_mat are not allowed (REQUIRED).}

\item{env}{numeric vector, dist, or square matrix. Environmental variable 
corresponding to abunds. For example, temperature, or geographic distance.
Not required for computing phylogenetic specificity (default: NULL).}

\item{hosts}{character vector. Host identities corresponding to abunds. Only
required if calculating phylogenetic specificity (default: NULL).}

\item{hosts_phylo}{phylo object. Tree containing all unique hosts as tips. Only
required if calculating phylogenetic specificity (default: NULL).}

\item{n_sim}{integer. Number of simulations of abunds_mat to do under the null 
hypothesis that host or environmental association is random. P-values will not
be calculated if n_sim < 100 (default: 500).}

\item{p_adj}{string. Type of multiple hypothesis testing correction performed
on P-values. Can take any valid method argument to p.adjust, including "none",
"bonferroni", "holm", "fdr", and others (default: "fdr").}

\item{seed}{integer. Seed to use so that this is repeatable. Same seed will be
used for each species in abunds_mat, so all species will experience the same 
permutations. This can be disabled by setting seed=0, which will make permutation
is both non deterministic (not repeatable) AND each species will experience
different permutations (default: 1234557).}

\item{tails}{integer. 1 = 1-tailed, test for specificity only. 2 = 2-tailed.
3 = 1-tailed, test for cosmopolitanism only. 0 = no test, P=1.0 (default: 1).}

\item{n_cores}{integer. Number of CPU cores to use for parallel operations. If
set to 1, lapply will be used instead of mclapply (default: 2).}

\item{verbose}{logical. Should status messages be displayed? (default: TRUE).}

\item{p_method}{string. "raw" for quantile method, or "gamma_fit" for calculating P
by fitting a gamma distribution (default: "raw").}

\item{center}{string. Type of central tendency to use for simulated RQE values.
Options are "mean", "median", and "mode". If mode is chosen, a reversible gamma
distribution is fit and mode is calculated using that distribution (default: mean).}

\item{denom_type}{string. Type of denominator (d) to use (default: "index"). Note
that denominator type does NOT affect P-values.
\describe{
  \item{"ses":}{
    d for species s is calculated as the standard deviation of RQE values
    calculated from permuted species weights. This makes the output specificity
    a standardized effect size (SES). Unfortunately, this makes SES 
    counterintuitively sensitive to occupancy, where species with high occupancy
    have more extreme SES than rare species, due to their more deterministic sim
    specificities. Included for comparative purposes, not suggested.
  }
  \item{"raw":}{
    d is 1 for all species, so output specificity has units of distance, i.e. the
    raw difference between empirical and simulated RQE. This means that results
    from different variables are not comparable, since it is not scale-invariant to
    env or hosts_phylo. It not scale-invariant to the species weights in aunds_mat,
    either. Not sensitive to number of samples. Not suggested because units are
    strange, and isn't comparable between variables. 
  }
  \item{"index":}{
    d is the mean of simulated (permuted) RQE values for species that have stronger
    specificity than expected by chance, resulting in specificity values with range
    [-1, 0), with 0 as the null hypothesis. In this case, -1 indicates perfect
    specificity, where a species is associated with zero environmental variability.
    In the euclidean sense, this could be a species that is always found at the
    exact same elevation or the exact same pH. For species that have weaker specificity
    than expected by chance, d is x minus the center (see above) of simulated RQE 
    values, where x is the maximum possible dissimilarity observable given species
    weights. This d has other useful properties: scale invariance to env/hosts_phylo,
    insensitivity to the number of samples, insensitivity to occupancy, and strong 
    sensitivity to specificity (default).
  }
}}

\item{diagnostic}{logical. If true, changes output to include different parts of Spec. 
This includes Pval, Spec, raw, denom, emp, and all sim values with column labels as
simN where N is the number of sims (default: FALSE)}
}
\value{
data.frame where each row is an input species. First column is P-value
  ($Pval), second column is specificity ($Spec).
}
\description{
Calculates species' specificities to either a 1-dimensional variable (vector), 
2-dimensional variable (matrix), or to a phylogeny. Transforms all variable
input types into a matrix D, and calculates specificity by comparing empirical
Rao's Quadratic Entropy to simulated RQE (same but with permuted abundances). 
By default (denom_type = "index"), an index is calculated from emp and sim
values such that Spec=0 indicates random assortment (null hypothesis), and more
negative values indicate stronger specificity.
}
\examples{
library(specificity)
attach(endophyte)
# only analyze species with occupancy >= 20
m <- occ_threshold(prop_abund(otutable), 20)
# create list to hold phy_or_env_spec outputs
specs_list <- list()

# phylogenetic specificity using endophyte data set
specs_list$host <- phy_or_env_spec(
    abunds_mat=m,
    hosts=metadata$PlantGenus, 
    hosts_phylo=supertree,
    n_sim=100, p_method="gamma_fit",
    n_cores=4
)

# environmental specificity using elevation from endophyte data set:
specs_list$elev <- phy_or_env_spec(
    abunds_mat=m,
    env=metadata$Elevation,
    n_sim=100, p_method="gamma_fit",
    n_cores=4
)

# geographic specificity using spatial data from endophyte data set:
specs_list$geo <- phy_or_env_spec(
    abunds_mat=m,
    env=distcalc(metadata$Lat, metadata$Lon),
    n_sim=100, p_method="gamma_fit",
    n_cores=4
)

plot_specs_violin(specs_list, cols=c("forestgreen", "red", "black"))

}
\references{
\itemize{
  \item Poulin et al. (2011) Host specificity in phylogenetic and geographic
    space. Trends Parasitol 8:355-361. doi: 10.1016/j.pt.2011.05.003
  \item Rao CR (2010) Quadratic entropy and analysis of diversity. Sankhya 
    72:70-80. doi: 10.1007/s13171-010-0016-3
  \item Rao CR (1982) Diversity and dissimilarity measurements: A unified 
    approach. Theor Popul Biol 21:24-43.
}
}
\author{
John L. Darcy
}
