% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_from_perms.r
\name{pval_from_perms}
\alias{pval_from_perms}
\title{pval_from_perms}
\usage{
pval_from_perms(
  emp,
  perm,
  tails,
  method = "raw",
  threshold = 30,
  rounding = -1
)
}
\arguments{
\item{emp}{Numeric scalar. An empirical test statistic value.}

\item{perm}{Numeric vector. Test statistic values similar to emp, but calculated
from permuted data.}

\item{tails}{integer. 
\describe{
  \item{1:}{Left tail only.}
  \item{2:}{2-tailed test.}
  \item{3:}{Right tail only.}
  \item{0:}{No test, P=1.}
}}

\item{method}{string. Method by which P should be calculated from perms:
\describe{
  \item{"raw":}{
    P is calculated as the sum of sim values more extreme than the empirical
    value plus one, divided by the number of sim values. 
  }
  \item{"dens_fit":}{
    P is calculated via kernel density estimation. No better than "raw".
  }
  \item{"gamma_fit":}{
    P is calculated by fitting a gamma distribution to sim values and calculating 
    area under the curve from (-inf,emp] or [emp,inf) depending on tailedness.
  }
}}

\item{threshold}{integer. Minimum number n of non-NA values in perm that are
acceptable. If n < threshold, P=NA (DEFAULT: 50).}

\item{rounding}{integer. Number of decimal places to round emp and perm This is only
useful when emp and perm are expected to contain the exact same value, but the number
of decimal places in that value is different between emp and perm Use a number less
then zero to disable rounding (DEFAULT: -1).}
}
\value{
a P-value.
}
\description{
Calculates P-value for permutation tests.
}
\author{
John L. Darcy
}
