% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rep_positions.r
\name{random_rep_positions}
\alias{random_rep_positions}
\title{random_rep_positions}
\usage{
random_rep_positions(x, nbins = 50)
}
\arguments{
\item{x}{vector}

\item{nbins}{number of bins to use}
}
\value{
integer vector of positions that were selected
}
\description{
Finds positions in a vector (or matrix) that are randomly located within n_bins
evenly sized bins. This is useful for 1:1 comparisons of large vectors where
plotting or comparing all points is prohibitive. Only used in an example for the
prop_abund() function.
}
\examples{
# library(specificity)
# attach(endophyte)
# m_dbl <- prop_abund(otutable)
# m_int <- prop_abund(otutable, to_int=TRUE)
# head(rowSums(m_dbl))
# head(rowSums(m_int))
# # note that they are off by a little bit. This small loss in precision is OK.
# object.size(m_dbl)
# object.size(m_int)
# random_positions <- random_rep_positions(m_dbl, 100)
# plot(m_int[random_positions] ~ m_dbl[random_positions])

}
\author{
John L. Darcy
}
