% NOT generated by roxygen2, go ahead and edit by hand
\name{raoperms}
\alias{raoperms}
\title{raoperms}
\usage{
raoperms(p,D,nsim,seed)
}
\arguments{
\item{p}{numeric vector of species weights.}
\item{D}{numeric vector (dist) of distances corresponding to a lower triangle of a matrix whose rows and cols correspond to p; i.e. an lxl matrix where l is length(p). R's dist() function does this for you!}
\item{nsim}{integer. number of sims to do}
\item{seed}{integer. seed for randomization}
}
\value{
Vector of Rao values, length = nsim.
}
\description{
C++ function used by phy_or_env_spec(). Not meant for use otherwise.
}
\author{
John L. Darcy
}