% NOT generated by roxygen2, go ahead and edit by hand
\name{raoperms}
\alias{raoperms}
\title{raoperms}
\usage{
raoperms(
	p,
	D,
	n_sim=1000,
	seed=12345)
}
\arguments{
\item{p}{numeric vector of species weights.}
\item{D}{numeric vector (dist) of distances corresponding to a lower triangle of a matrix whose rows and cols correspond to p; i.e. an lxl matrix where l is length(p). R's dist() function does this for you!}
\item{n_sim}{integer. number of sims to do (default: 1000).}
\item{seed}{integer. a seed to be used for permutation. setting seed=0 will give nondeterministic random results, as if no seed were set (default: 0).}
}
\value{
Vector of Rao values, length = n_sim.
}
\description{
C++ function used by phy_or_env_spec(). Not meant for use otherwise.
}
\examples{
  library(specificity)
  p <- runif(100) 
  D <- dist(runif(100))
  a <- raoperms(p,D,100,12345)
  hist(a)
}

\author{
John L. Darcy
}