% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2mat.r
\name{tree2mat}
\alias{tree2mat}
\title{tree2mat}
\usage{
tree2mat(tree, x, n_cores = 1, delim = ";")
}
\arguments{
\item{tree}{phylo object. Tree containing all unique species in x as tips. May
contain tips that are not in x.}

\item{x}{character vector. Vector of species identities, each of which must
be in tree as a tip label. May contain any given species identity more than
once.}

\item{n_cores}{integer. Number of cores to use for parallel computation. No
parallelization will be done if n_cores = 1. Multithreading should only be used
for large trees where x has low redundancy (default: 1).}

\item{delim}{string. Delimiter character or string for internal use. Must not be
present in tree$tip.label. This is checked by the function and will return an
error otherwise (default: ";").}
}
\value{
dist object, of vector length equal to (l^2-l)/2 where l is length(x); i.e. 
  values are the lower triangle of a patristic distance matrix with rows=x and cols=x.
}
\description{
Transforms a phylogenetic tree into a dist object containing patristic distances
between tips. Dists are just lower triangles of matrices, and the rows and
columns of that matrix are defined by a user-supplied vector of tip labels,
which can include duplicate values. Contrast with ape::cophenetic.phylo, which
produces a distance matrix containing only unique pairwise patristic distances
within the phylogeny.
}
\examples{
  library(specificity)
  library(ape)
  example_tree <- ape::read.tree(text=" ((((a:1,b:1):1,c:2):1,d:3):1,(e:1,f:1):3);")
  example_x <- c("a", "a", "a", "b", "c", "d", "c", "a", "f")
  # unique patristic distance matrix:
  ape::cophenetic.phylo(example_tree)
  # dist object for example_x:
  tree2mat(tree=example_tree, x=example_x)

  # examples with other delimiters
  tree2mat(tree=example_tree, x=example_x, delim="@")
  tree2mat(tree=example_tree, x=example_x, delim="i love cats")
  # should fail since "a" is in a tip name:
  # tree2mat(tree=example_tree, x=example_x, delim="a")

}
\author{
John L. Darcy
}
