% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10__wpd.r
\name{wpd}
\alias{wpd}
\title{wpd}
\usage{
wpd(s, s_phylo, w = NULL, nested_set = NULL, metric = "Hp")
}
\arguments{
\item{s}{character vector. One species name per observation. If no species was
observed for a given datum, use NA. s can also be provided as a vector of
unique species identities, in which case counts of those species can be
given as w.}

\item{s_phylo}{phylo object. Tree containing all unique names in s as tips.
Must not contain duplicate tip labels.}

\item{w}{numeric vector. Optional weights for s, e.g. number of parasites 
observed in each sample, or boolean weights corresponding to presence or 
absence of parasite species, or confidence species was observed, etc. If w is
not provided but a weighted metric is specified, w will be set to 1 for each
value of s. Thus, weights for each unique species in s would be equal to the
number of times that species appears in s. w is not used for unweighted
metrics (PD). Any NA values in w will be pairwise removed from w and s
(DEFAULT: NULL).}

\item{nested_set}{matrix. The output of make_nested_set(s_phylo). If not
provided, will be calculated on the fly. Precalculation only provides speedup
with very large trees (DEFAULT: NULL).}

\item{metric}{character. Abbreviated name of desired tree-based phylogenetic
diversity metric. Available metrics are:
\describe{
  \item{Hp:}{
    Phylogenetic Entropy. Insensitive to 0 weights, cannot increase with removal
    of taxa. Allen et al. 2009.
  }
  \item{WF:}{
    Weighted Faith's PD. Sensitive to 0 weights, i.e. a clade that was heavily
    sampled but has lots of zeroes will cause its sister clades to be
    underrepresented. Swenson 2014.
  }
  \item{PD:}{
    Original Faith's Phylogenetic Diversity. Unweighted. Simply a sum of branch-
    lengths in your tree (but only for taxa in s). Faith 1992.
  }
}}
}
\value{
Single WPD or PD value.
}
\description{
Calculates weighted Phylogenetic Diversity for a vector s of species
observations, weighted by the frequency of each species within s. For example,
if S={a, a, b, a, b, c, a}, then species a will have weight 4, species b will
have weight 2, and species c will have weight 1. Unobserved species have weight
zero.
}
\details{
However, one may wish to exclude observations that do not meet some criterion,
such as co-observation of a symbiote or parasite. For this reason, a second set
of weights w can be provided as a vector of numeric values that are paired with
s. These weights are then implicitely combined with the weights discussed above 
depending on which weighted metric is chosen. In the case of Phylogenetic 
Entropy (Hw), per-tip weights are calculated as the sums of w. In the case of
Weighted Faith (WF), per-tip weights are averages of w.
}
\examples{
  library(geiger)
  set.seed(12345)
  s_phylo <- get(data(geospiza))$phy
  w <- sample(c(0, 1), replace=T, size=10)
  s <- sample(s_phylo$tip.label, replace=T, size=10)
  wpd(s, s_phylo, w, metric="Hp")

}
\references{
Allen B, Kon M, Bar-Yam Y (2009) A new phylogenetic diversity measure
  generalizing the Shannon index and its application to Phyllostomid bats.
  American Naturalist 174(2).
Swenson NG (2014) Functional and Phylogenetic Ecology in R. 
  Springer UseR! Series, Springer, New York, New York, U.S.A.
Faith DP (1992) Conservation evaluation and phylogenetic diversity. Biological
  Conservation 61.
}
\seealso{
rao_quad_ent, a phylogenetic diversity measure that uses a distance matrix
  instead of a phylogenetic tree.
}
\author{
John L. Darcy
}
